Update preprocessing test.ipynb
this commit is broken. says "Int is not a callable object" when running __len__(). initially, I was trying to check if, for some reason, all the labels were set to one value. I wanted to loopthrough the labels but now I can't do that. I cant even define the dataset